Add files via upload
Titanic Dataset Cleaning Description
Cleaning the Titanic dataset is a crucial part of ensuring the data is ready for analysis or use in machine learning models. The main steps performed in dataset cleaning are detailed below.

1. Duplicate Removal
Duplicate rows were removed to avoid bias and ensure that each sample in the dataset is unique. This helps prevent models from learning erroneous patterns due to repeated data.

2. Handling Missing Values
Age: The age column contains missing values. These were filled with the median of the column, as the median is less sensitive to extreme values ​​than the mean, providing a robust solution.

Embarked: Some records have missing values ​​in this column. Missing values ​​were replaced with the most frequent value (mode) in the column, as this value best represents the current data.

Cabin: This column has many missing values, so it was decided to replace them with the value "Unknown," as this could be useful for indicating the absence of information.

3. Converting Categorical Variables
To allow machine learning models to work with the data effectively, categorical variables were converted to numeric values:

The Sex column was transformed, assigning 0 to "male" and 1 to "female."

The Embarked port column was converted to numeric values, assigning 0, 1, and 2 to ports "S," "C," and "Q," respectively.